Fix some format string mismatches (int versus long int).

+++ nsock/src/engine_select.c
@@ -78,7 +78,7 @@
     } else { \
       fatal("%s:%ld: Attempt to FD_SET fd %d, which is not less than" \
         " FD_SETSIZE (%d). Try using a lower parallelism.", \
-        __FILE__, __LINE__, (fd), FD_SETSIZE); \
+        __FILE__, (long int) __LINE__, (fd), FD_SETSIZE); \
     } \
   } while (0)
 #endif
@@ -93,7 +93,7 @@
     } else { \
       fatal("%s:%ld: Attempt to FD_CLR fd %d, which is not less than" \
         " FD_SETSIZE (%d). Try using a lower parallelism.", \
-        __FILE__, __LINE__, (fd), FD_SETSIZE); \
+        __FILE__, (long int) __LINE__, (fd), FD_SETSIZE); \
     } \
   } while (0)
 #endif
